add several spawn positions for enemies in waves
