+++ Assets/Scripts/Missions/WavesController.cs
@@ -1,4 +1,5 @@
 using System.Collections;
+using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Events;
 
@@ -7,8 +8,7 @@ public enum SpawnSide
     Left,
     Right,
     Top,
-    Bottom,
-    All
+    Bottom
 }
 
 public struct Wave
@@ -19,7 +19,6 @@ public struct Wave
 
 public class WavesController : MonoBehaviour
 {
-
     private Enemy enemyController;
 
     [SerializeField]
@@ -27,7 +26,7 @@ public class WavesController : MonoBehaviour
 
     [SerializeField]
     private float waveDelayInSeconds = 10.0f;
-
+    
     [SerializeField]
     private int waveCount = 3;
 
@@ -39,6 +38,8 @@ public class WavesController : MonoBehaviour
 
     private int currentWaveIndex = 0;
 
+    private readonly Dictionary<SpawnSide, Transform> spawnPositions = new();
+
     public event UnityAction<float> WaveAwaited;
     public event UnityAction WaveStarted;
     public event UnityAction WaveEnded;
@@ -47,7 +48,13 @@ public class WavesController : MonoBehaviour
     private void Start()
     {
         enemyController = GetComponent<Enemy>();
+
         StartCoroutine(StartMissionWavesRoutine());
+
+        spawnPositions[SpawnSide.Left]   = transform.Find("Left Position");
+        spawnPositions[SpawnSide.Right]  = transform.Find("Right Position");
+        spawnPositions[SpawnSide.Top]    = transform.Find("Top Position");
+        spawnPositions[SpawnSide.Bottom] = transform.Find("Bottom Position");
     }
 
     private IEnumerator StartMissionWavesRoutine()
@@ -59,9 +66,12 @@ public class WavesController : MonoBehaviour
     private IEnumerator StartWaveRoutine(float delayInSeconds)
     {
         WaveAwaited?.Invoke(delayInSeconds);
+
         yield return AwaitWaveRoutine(delayInSeconds);
         
         WaveStarted?.Invoke();
+        enemyController.ViewRadius = 1000.0f;
+
         var enemyCount = enemiesPerWave[currentWaveIndex % enemiesPerWave.Length];
         var spawnSide = spawnSides[currentWaveIndex % spawnSides.Length];
         yield return SpawnEnemiesRoutine(enemyCount, spawnSide);
@@ -74,12 +84,13 @@ public class WavesController : MonoBehaviour
 
         enemyController.UnitDied += OnEnemyUnitDied;
 
+        var spawnPosition = spawnPositions[side].position;
+        var positions = PositionUtils.GetPositionsAround(spawnPosition, enemyCount);
+
         for (var i = 0; i < enemyCount; i++)
         {
-            // TODO: select position smarter
-            var position = Vector2.zero;
+            var position = positions[i % positions.Count];
             enemyController.SpawnEnemyUnit(position);
-            yield return new WaitForSeconds(1.0f);
         }
     }
 
@@ -89,6 +100,7 @@ public class WavesController : MonoBehaviour
             return;
 
         WaveEnded?.Invoke();
+        enemyController.ViewRadius = 10.0f;
 
         enemyController.UnitDied -= OnEnemyUnitDied;
         if (currentWaveIndex + 1 >= waveCount)
